Factor types and utils to separate files
